requirements.txt
# urlRecon — Python 3.10+ dependencies

# WHOIS / ASN lookup
ipwhois

# DNS queries
dnspython

# KML file generation
simplekml

# HTTP client
requests
